# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * NutritionGhar API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface SimpleSuccess {
  success: boolean;
}

export type UserRole = (typeof UserRole)[keyof typeof UserRole];

export const UserRole = {
  customer: "customer",
  chef: "chef",
} as const;

export type DishCategory = (typeof DishCategory)[keyof typeof DishCategory];

export const DishCategory = {
  veg: "veg",
  non_veg: "non_veg",
  diet: "diet",
  gym: "gym",
  elderly: "elderly",
  tiffin: "tiffin",
} as const;

export type DishTag = (typeof DishTag)[keyof typeof DishTag];

export const DishTag = {
  healthy: "healthy",
  protein: "protein",
  low_calorie: "low_calorie",
  tiffin: "tiffin",
} as const;

export type OrderStatus = (typeof OrderStatus)[keyof typeof OrderStatus];

export const OrderStatus = {
  placed: "placed",
  preparing: "preparing",
  out_for_delivery: "out_for_delivery",
  delivered: "delivered",
  cancelled: "cancelled",
} as const;

export type SubscriptionFrequency =
  (typeof SubscriptionFrequency)[keyof typeof SubscriptionFrequency];

export const SubscriptionFrequency = {
  weekly: "weekly",
  monthly: "monthly",
} as const;

export interface User {
  id: string;
  phone: string;
  name: string;
  role: UserRole;
  kitchenName?: string | null;
  bio?: string | null;
  address?: string | null;
  createdAt: string;
}

export interface SendOtpInput {
  /** @minLength 10 */
  phone: string;
}

export interface SendOtpResult {
  success: boolean;
  demoOtp: string;
}

export interface VerifyOtpInput {
  phone: string;
  otp: string;
  role: UserRole;
  name?: string;
}

export interface AuthResponse {
  user: User;
}

export interface MeResponse {
  user: User | null;
}

export interface UpdateProfileInput {
  name?: string;
  kitchenName?: string;
  bio?: string;
  address?: string;
}

export interface Dish {
  id: string;
  chefId: string;
  chefName: string;
  kitchenName?: string | null;
  name: string;
  description: string;
  /** Price in INR */
  price: number;
  category: DishCategory;
  tags: DishTag[];
  prepMinutes: number;
  calories?: number | null;
  proteinGrams?: number | null;
  imageUrl?: string | null;
  rating: number;
  reviewCount: number;
  ecoFriendly: boolean;
  available: boolean;
  createdAt: string;
}

export interface CreateDishInput {
  name: string;
  description: string;
  price: number;
  category: DishCategory;
  tags?: DishTag[];
  prepMinutes: number;
  calories?: number | null;
  proteinGrams?: number | null;
  imageUrl?: string | null;
  ecoFriendly?: boolean;
  available?: boolean;
}

export interface UpdateDishInput {
  name?: string;
  description?: string;
  price?: number;
  category?: DishCategory;
  tags?: DishTag[];
  prepMinutes?: number;
  calories?: number | null;
  proteinGrams?: number | null;
  imageUrl?: string | null;
  ecoFriendly?: boolean;
  available?: boolean;
}

export interface OrderItem {
  dishId: string;
  dishName: string;
  quantity: number;
  unitPrice: number;
}

export interface Order {
  id: string;
  customerId: string;
  customerName: string;
  customerPhone: string;
  chefId: string;
  chefName: string;
  kitchenName?: string | null;
  items: OrderItem[];
  subtotal: number;
  deliveryFee: number;
  total: number;
  status: OrderStatus;
  address: string;
  notes?: string | null;
  estimatedMinutes: number;
  placedAt: string;
  scheduledFor?: string | null;
  isCorporate: boolean;
}

export interface CreateOrderItem {
  dishId: string;
  /** @minimum 1 */
  quantity: number;
}

export interface CreateOrderInput {
  items: CreateOrderItem[];
  address: string;
  notes?: string;
  isCorporate?: boolean;
  scheduledFor?: string;
}

export interface UpdateOrderStatusInput {
  status: OrderStatus;
}

export interface SubscriptionPlan {
  id: string;
  name: string;
  description: string;
  pricePerMeal: number;
  mealsPerWeek: number;
  frequency: SubscriptionFrequency;
  category: DishCategory;
  forCorporate: boolean;
  highlights: string[];
}

export interface Subscription {
  id: string;
  userId: string;
  planId: string;
  planName: string;
  startDate: string;
  deliveryAddress: string;
  active: boolean;
}

export interface CreateSubscriptionInput {
  planId: string;
  deliveryAddress: string;
}

export interface ChefStats {
  totalOrders: number;
  activeOrders: number;
  deliveredOrders: number;
  revenue: number;
  avgRating: number;
  topDishes: Dish[];
}

export interface CategoryCount {
  category: DishCategory;
  count: number;
}

export interface HomeFeatured {
  featuredDishes: Dish[];
  topRatedDishes: Dish[];
  topChefs: User[];
  categoryCounts: CategoryCount[];
}

export type ListDishesParams = {
  q?: string;
  category?: DishCategory;
  tag?: DishTag;
  chefId?: string;
};