# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: NutritionGhar API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/send-otp:
    post:
      operationId: sendOtp
      tags: [auth]
      summary: Send a one-time password to the phone number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendOtpInput"
      responses:
        "200":
          description: OTP issued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SendOtpResult"
  /auth/verify-otp:
    post:
      operationId: verifyOtp
      tags: [auth]
      summary: Verify OTP and start a session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerifyOtpInput"
      responses:
        "200":
          description: Session established
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
...
  /auth/profile:
    patch:
      operationId: updateProfile
      tags: [auth]
      summary: Update current user's profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProfileInput"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /dishes:
    get:
      operationId: listDishes
      tags: [dishes]
      summary: List dishes (with optional search and filter)
      parameters:
        - name: q
          in: query
          required: false
          schema: { type: string }
        - name: category
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/DishCategory"
        - name: tag
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/DishTag"
        - name: chefId
          in: query
          required: false
          schema: { type: string }
      responses:
        "200":
          description: Dish list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Dish"
  /dishes/{id}:
    get:
      operationId: getDish
      tags: [dishes]
      summary: Get one dish
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Dish
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Dish"
  /chef/dishes:
    get:
      operationId: listChefDishes
      tags: [chef]
      summary: List dishes owned by the current chef
      responses:
        "200":
          description: Chef dish list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Dish"
    post:
      operationId: createDish
      tags: [chef]
      summary: Create a dish
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDishInput"
      responses:
        "200":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Dish"
  /chef/dishes/{id}:
    patch:
      operationId: updateDish
      tags: [chef]
      summary: Update a dish
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateDishInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Dish"
    delete:
      operationId: deleteDish
      tags: [chef]
      summary: Delete a dish
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SimpleSuccess"
...
  /chef/orders/{id}/status:
    patch:
      operationId: updateOrderStatus
      tags: [chef]
      summary: Advance the status of one chef order
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateOrderStatusInput"
      responses:
        "200":
          description: Updated order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
...
  /orders:
    get:
      operationId: listMyOrders
      tags: [orders]
      summary: List the current customer's orders
      responses:
        "200":
          description: Order list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
    post:
      operationId: createOrder
      tags: [orders]
      summary: Place a new order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrderInput"
      responses:
        "200":
          description: Created order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get one order (owner or assigned chef)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
...
components:
  schemas:
...
    SimpleSuccess:
      type: object
      properties:
        success: { type: boolean }
      required: [success]
    UserRole:
      type: string
      enum: [customer, chef]
    DishCategory:
      type: string
      enum: [veg, non_veg, diet, gym, elderly, tiffin]
    DishTag:
      type: string
      enum: [healthy, protein, low_calorie, tiffin]
    OrderStatus:
      type: string
      enum: [placed, preparing, out_for_delivery, delivered, cancelled]
...
    User:
      type: object
      properties:
        id: { type: string }
        phone: { type: string }
        name: { type: string }
        role:
          $ref: "#/components/schemas/UserRole"
        kitchenName: { type: string, nullable: true }
        bio: { type: string, nullable: true }
        address: { type: string, nullable: true }
        createdAt: { type: string, format: date-time }
      required: [id, phone, name, role, createdAt]
    SendOtpInput:
      type: object
      properties:
        phone: { type: string, minLength: 10 }
      required: [phone]
    SendOtpResult:
      type: object
      properties:
        success: { type: boolean }
        # Mock OTP for the demo - returned to client so users can see it
        demoOtp: { type: string }
      required: [success, demoOtp]
    VerifyOtpInput:
      type: object
      properties:
        phone: { type: string }
        otp: { type: string }
        role:
          $ref: "#/components/schemas/UserRole"
        name: { type: string }
      required: [phone, otp, role]
    AuthResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
      required: [user]
...
    UpdateProfileInput:
      type: object
      properties:
        name: { type: string }
        kitchenName: { type: string }
        bio: { type: string }
        address: { type: string }
    Dish:
      type: object
      properties:
        id: { type: string }
        chefId: { type: string }
        chefName: { type: string }
        kitchenName: { type: string, nullable: true }
        name: { type: string }
        description: { type: string }
        price:
          type: number
          description: Price in INR
        category:
          $ref: "#/components/schemas/DishCategory"
        tags:
          type: array
          items:
            $ref: "#/components/schemas/DishTag"
        prepMinutes: { type: integer }
        calories: { type: integer, nullable: true }
        proteinGrams: { type: integer, nullable: true }
        imageUrl: { type: string, nullable: true }
        rating: { type: number }
        reviewCount: { type: integer }
        ecoFriendly: { type: boolean }
        available: { type: boolean }
        createdAt: { type: string, format: date-time }
      required:
        - id
        - chefId
        - chefName
        - name
        - description
        - price
        - category
        - tags
        - prepMinutes
        - rating
        - reviewCount
        - ecoFriendly
        - available
        - createdAt
    CreateDishInput:
      type: object
      properties:
        name: { type: string }
        description: { type: string }
        price: { type: number }
        category:
          $ref: "#/components/schemas/DishCategory"
        tags:
          type: array
          items:
            $ref: "#/components/schemas/DishTag"
        prepMinutes: { type: integer }
        calories: { type: integer, nullable: true }
        proteinGrams: { type: integer, nullable: true }
        imageUrl: { type: string, nullable: true }
        ecoFriendly: { type: boolean }
        available: { type: boolean }
      required: [name, description, price, category, prepMinutes]
    UpdateDishInput:
      type: object
      properties:
        name: { type: string }
        description: { type: string }
        price: { type: number }
        category:
          $ref: "#/components/schemas/DishCategory"
        tags:
          type: array
          items:
            $ref: "#/components/schemas/DishTag"
        prepMinutes: { type: integer }
        calories: { type: integer, nullable: true }
        proteinGrams: { type: integer, nullable: true }
        imageUrl: { type: string, nullable: true }
        ecoFriendly: { type: boolean }
        available: { type: boolean }
    OrderItem:
      type: object
      properties:
        dishId: { type: string }
        dishName: { type: string }
        quantity: { type: integer }
        unitPrice: { type: number }
      required: [dishId, dishName, quantity, unitPrice]
    Order:
      type: object
      properties:
        id: { type: string }
        customerId: { type: string }
        customerName: { type: string }
        customerPhone: { type: string }
        chefId: { type: string }
        chefName: { type: string }
        kitchenName: { type: string, nullable: true }
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        subtotal: { type: number }
        deliveryFee: { type: number }
        total: { type: number }
        status:
          $ref: "#/components/schemas/OrderStatus"
        address: { type: string }
        notes: { type: string, nullable: true }
        estimatedMinutes: { type: integer }
        placedAt: { type: string, format: date-time }
        scheduledFor: { type: string, format: date-time, nullable: true }
        isCorporate: { type: boolean }
      required:
        - id
        - customerId
        - customerName
        - customerPhone
        - chefId
        - chefName
        - items
        - subtotal
        - deliveryFee
        - total
        - status
        - address
        - estimatedMinutes
        - placedAt
        - isCorporate
    CreateOrderItem:
      type: object
      properties:
        dishId: { type: string }
        quantity: { type: integer, minimum: 1 }
      required: [dishId, quantity]
    CreateOrderInput:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/CreateOrderItem"
        address: { type: string }
        notes: { type: string }
        isCorporate: { type: boolean }
        scheduledFor: { type: string, format: date-time }
      required: [items, address]
    UpdateOrderStatusInput:
      type: object
      properties:
        status:
          $ref: "#/components/schemas/OrderStatus"
      required: [status]
...