# lib/api-zod/src/generated/types/dishCategory.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * NutritionGhar API
 * OpenAPI spec version: 0.1.0
 */

export type DishCategory = (typeof DishCategory)[keyof typeof DishCategory];

export const DishCategory = {
  veg: "veg",
  non_veg: "non_veg",
  diet: "diet",
  gym: "gym",
  elderly: "elderly",
  tiffin: "tiffin",
} as const;