# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * NutritionGhar API
 * OpenAPI spec version: 0.1.0
 */
import type { OrderItem } from "./orderItem";
import type { OrderStatus } from "./orderStatus";

export interface Order {
  id: string;
  customerId: string;
  customerName: string;
  customerPhone: string;
  chefId: string;
  chefName: string;
  kitchenName?: string | null;
  items: OrderItem[];
  subtotal: number;
  deliveryFee: number;
  total: number;
  status: OrderStatus;
  address: string;
  notes?: string | null;
  estimatedMinutes: number;
  placedAt: Date;
  scheduledFor?: Date | null;
  isCorporate: boolean;
}